test: split tests into per-package files
Splits test_package_addons.py into per-package test files. Also
splits the test_versions function in test_dist_packages_versions.py
into three per-package functions. After both splits each test
depends on at most one package, which makes it possible to mark
each with a single requires_option in a follow-up commit and skip
it cleanly when its package was not built.

Ticket: MEN-9667
