+++ tests/test_package_monitor.py
@@ -1,5 +1,5 @@
 #!/usr/bin/python3
-# Copyright 2022 Northern.tech AS
+# Copyright 2026 Northern.tech AS
 #
 #    Licensed under the Apache License, Version 2.0 (the "License");
 #    you may not use this file except in compliance with the License.
@@ -14,72 +14,12 @@
 #    limitations under the License.
 
 import pytest
-import re
 
 from helpers import package_filename, upload_deb_package, check_installed
 
 
 @pytest.mark.usefixtures("setup_mender_configured")
-class TestPackageAddons:
-    def test_mender_connect(
-        self,
-        setup_tester_ssh_connection,
-        mender_dist_packages_versions,
-        mender_connect_version,
-    ):
-        # Upload
-        upload_deb_package(
-            setup_tester_ssh_connection,
-            mender_dist_packages_versions["mender-connect"],
-            "mender-connect",
-        )
-
-        # Install
-        setup_tester_ssh_connection.run(
-            "sudo apt -y install --fix-broken ./"
-            + package_filename(
-                mender_dist_packages_versions["mender-connect"], "mender-connect"
-            )
-        )
-        check_installed(setup_tester_ssh_connection, "mender-connect")
-
-        # Check mender-connect files
-        setup_tester_ssh_connection.run("test -x /usr/bin/mender-connect")
-        setup_tester_ssh_connection.run("test -f /etc/mender/mender-connect.conf")
-        setup_tester_ssh_connection.run(
-            "test -f /lib/systemd/system/mender-connect.service"
-        )
-
-    def test_mender_configure(
-        self, setup_tester_ssh_connection, mender_dist_packages_versions
-    ):
-        # Upload
-        upload_deb_package(
-            setup_tester_ssh_connection,
-            mender_dist_packages_versions["mender-configure"],
-            "mender-configure",
-            "all",
-        )
-
-        # Install
-        setup_tester_ssh_connection.run(
-            "sudo apt -y install --fix-broken ./"
-            + package_filename(
-                mender_dist_packages_versions["mender-configure"],
-                "mender-configure",
-                "all",
-            )
-        )
-        check_installed(setup_tester_ssh_connection, "mender-configure")
-
-        # Check mender-configure files
-        setup_tester_ssh_connection.run(
-            "test -x /usr/share/mender/modules/v3/mender-configure"
-        )
-        setup_tester_ssh_connection.run(
-            "test -x /usr/share/mender/inventory/mender-inventory-mender-configure"
-        )
-
+class TestPackageMonitor:
     @pytest.mark.commercial
     def test_mender_monitor(
         self, setup_tester_ssh_connection, mender_dist_packages_versions
